pyproject.toml
[project]
name = "transport-rental"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "email-validator>=2.2.0",
    "fakeredis>=2.30.1",
    "fastapi>=0.116.1",
    "fastapi-limiter>=0.1.6",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "starlette-exporter>=0.23.0",
    "uvicorn>=0.35.0",
]
